[refactoring] WebServer, Main 수정 - application 패키지를 프로그래머가 작성하는 곳, webserver는 was에 관한 코드 보관소로 분리
