+++ src/main/java/codesquad/application/Main.java
@@ -1,0 +1,11 @@
+package codesquad.application;
+
+
+import codesquad.webserver.server.WebServer;
+
+public class Main {
+    public static void main(String[] args) {
+        WebServer webServer = new WebServer();
+        webServer.run();
+    }
+}
